% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pal.R
\name{check_pal}
\alias{check_pal}
\title{Show colours in gov.uk colour palette}
\usage{
check_pal(x = gov_cols)
}
\arguments{
\item{x}{A vector of colours. Defaults to \code{\link{gov_cols}}.}
}
\value{
A bar chart displaying the gov.uk approved colour scheme.
}
\description{
\code{\link{check_pal}} produces a bar chart showing the
gov.uk approved colours as available \url{here} and in the vector
\code{\link{gov_cols}}.
}
\examples{

\dontrun{
library(govstyle)

check_pal()
   }

}

