% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gov_cols.R
\docType{data}
\name{gov_cols}
\alias{gov_cols}
\title{A vector of the gov.uk approved colours}
\format{An object of class \code{character} of length 14.}
\usage{
gov_cols
}
\value{
A list of named hexadecimal colours.
}
\description{
\code{\link{gov_cols}} Provides a vector of named colours
from the approved gov.uk colour list available \url{here}.
}
\examples{

library(ggplot2)
library(govstyle)
library(dplyr)

mtcars \%>\%
   ggplot +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(
   colour = gov_cols["green"]
   ) +
   theme_gov()

}
\keyword{datasets}

