% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVA.R
\docType{data}
\name{GVA}
\alias{GVA}
\title{GVA data for 2010:2015.}
\format{A tibble with 12 rows and 3 variables: \describe{ \item{sector}{DCMS
  sector, one of: \code{ c('UK', 'all_sectors') }}, \item{year}{calendar
  year.} \item{GVA}{gross value added (GBP millions)}}.}
\source{
\url{https://www.gov.uk/government/statistics/dcms-sectors-economic-estimates-2016}
}
\usage{
GVA
}
\description{
Subset of data extracted from the 2016 Economic Estimates of DCMS Sectors 
2016 report. These data are provided for testing purposes.
}
\keyword{DCMS}
\keyword{Economic}
\keyword{Estimates}
\keyword{Sectors}
