% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pal.R
\name{check_pal}
\alias{check_pal}
\title{Show colours in gov.uk colour palette}
\usage{
check_pal(x = gov_cols)
}
\arguments{
\item{\code{x}}{Either an integer in which case: display \code{gov_cols(1:x)}
or a character or integer vector, in which case return \code{gov_cols(x)}.
Can also be specified manually: see examples.}
}
\value{
A bar chart displaying the gov.uk approved colour scheme.
}
\description{
\code{\link{check_pal}} produces a bar chart showing the
gov.uk approved colours as available \url{here} and in the vector
\code{\link{gov_cols}}.
}
\examples{


library(govstyle)

## Show a subset of gov.uk colours

## Choose n colours:

check_pal(2)

check_pal(4)

## Pick a range of colours

check_pal(3:8)

check_pal(c(1,10))

## Pick with character vector

check_pal(
x = gov_cols[c("green", "grass_green")]
)

}

